Add generic neighbor size.
This commit includes a generic neighbor size when generating a CSV file.
It also fixes a bug in CNN csv generation where it was depending on the
order which the ondes were visited for the surrounding violating nodes.
It also includes a pair of functions to save and calculated displacement
for two different implementations of a same design.